Migrate TopicController specifiec Test into its own Test class
