+++ src/test/java/com/hermesworld/ais/galapagos/topics/service/impl/TopicServiceImplTest.java
@@ -1,6 +1,8 @@
 package com.hermesworld.ais.galapagos.topics.service.impl;
 
-import com.hermesworld.ais.galapagos.applications.*;
+import com.hermesworld.ais.galapagos.applications.ApplicationMetadata;
+import com.hermesworld.ais.galapagos.applications.ApplicationsService;
+import com.hermesworld.ais.galapagos.applications.KnownApplication;
 import com.hermesworld.ais.galapagos.applications.impl.KnownApplicationImpl;
 import com.hermesworld.ais.galapagos.events.GalapagosEventManagerMock;
 import com.hermesworld.ais.galapagos.kafka.KafkaCluster;
@@ -13,8 +15,6 @@ import com.hermesworld.ais.galapagos.subscriptions.SubscriptionMetadata;
 import com.hermesworld.ais.galapagos.subscriptions.service.SubscriptionService;
 import com.hermesworld.ais.galapagos.topics.*;
 import com.hermesworld.ais.galapagos.topics.config.GalapagosTopicConfig;
-import com.hermesworld.ais.galapagos.topics.controller.TopicController;
-import com.hermesworld.ais.galapagos.topics.controller.UpdateTopicDto;
 import com.hermesworld.ais.galapagos.topics.service.ValidatingTopicService;
 import com.hermesworld.ais.galapagos.util.FutureUtil;
 import org.json.JSONObject;
@@ -770,43 +770,6 @@ public class TopicServiceImplTest {
 
         assertEquals("this topic is now a nice one :)", savedTopic.getDescription());
 
-    }
-
-    @Test
-    public void testDontResetDeprecationWhenTopicDescChanges() throws Exception {
-
-        SubscriptionService subscriptionService = mock(SubscriptionService.class);
-
-        TopicServiceImpl service = new TopicServiceImpl(kafkaClusters, applicationsService, topicNameValidator,
-                userService, topicConfig, eventManager);
-        ValidatingTopicServiceImpl validatingService = new ValidatingTopicServiceImpl(service, subscriptionService,
-                applicationsService, kafkaClusters, topicConfig);
-
-        UpdateTopicDto dto = new UpdateTopicDto("topic is now deprecated", LocalDate.of(2199, 2, 14),
-                "updated description goes here", true);
-
-        TopicMetadata topic = new TopicMetadata();
-        topic.setName("topic-1");
-        topic.setDeprecated(true);
-        topic.setEolDate(LocalDate.of(2299, 12, 4));
-        topic.setDescription("this topic is not a nice one :(");
-        topic.setOwnerApplicationId("app-1");
-        topic.setType(TopicType.EVENTS);
-        topicRepository.save(topic).get();
-
-        ApplicationOwnerRequest req = new ApplicationOwnerRequest();
-        req.setApplicationId("app-1");
-        req.setState(RequestState.APPROVED);
-
-        when(applicationsService.getUserApplicationOwnerRequests()).thenReturn((List.of(req)));
-
-        TopicController controller = new TopicController(validatingService, kafkaClusters, applicationsService, topicNameValidator);
-
-        controller.updateTopic("test", "topic-1", dto);
-        TopicMetadata savedTopic = topicRepository.getObject("topic-1").get();
-
-        assertTrue(savedTopic.isDeprecated());
-        assertEquals(topic.getEolDate(), savedTopic.getEolDate());
     }
 
     @Test
